+++ packages/dynamic-bonding-curve/src/helpers/validators.ts
@@ -1,5 +1,5 @@
 import BN from 'bn.js'
-import { MAX_CURVE_POINT, MAX_SQRT_PRICE, MIN_SQRT_PRICE } from './constants'
+import { MAX_CURVE_POINT, MAX_SQRT_PRICE, MIN_SQRT_PRICE } from '../constants'
 import {
     ActivationType,
     CollectFeeMode,
@@ -9,7 +9,7 @@ import {
     TokenType,
     type CreateConfigParam,
     type PoolConfig,
-} from './types'
+} from '../types'
 import { Connection, PublicKey } from '@solana/web3.js'
 import {
     getBaseTokenForSwap,
@@ -17,28 +17,11 @@ import {
     getMigrationThresholdPrice,
     getSwapAmountWithBuffer,
 } from './common'
-import { isNativeSol } from './utils'
-
-/**
- * Check if the locked vesting is the default
- * @param lockedVesting - The locked vesting parameters
- * @returns true if the locked vesting is the default, false otherwise
- */
-export function isDefaultLockedVesting(lockedVesting: {
-    amountPerPeriod: BN
-    cliffDurationFromMigrationTime: BN
-    frequency: BN
-    numberOfPeriod: BN
-    cliffUnlockAmount: BN
-}): boolean {
-    return (
-        lockedVesting.amountPerPeriod.eqn(0) &&
-        lockedVesting.cliffDurationFromMigrationTime.eqn(0) &&
-        lockedVesting.frequency.eqn(0) &&
-        lockedVesting.numberOfPeriod.eqn(0) &&
-        lockedVesting.cliffUnlockAmount.eqn(0)
-    )
-}
+import {
+    getTotalTokenSupply,
+    isDefaultLockedVesting,
+    isNativeSol,
+} from './utils'
 
 /**
  * Validate the pool fees
@@ -56,50 +39,6 @@ export function validatePoolFees(poolFees: any): boolean {
     }
 
     return true
-}
-
-/**
- * Get the total token supply
- * @param swapBaseAmount - The swap base amount
- * @param migrationBaseThreshold - The migration base threshold
- * @param lockedVestingParams - The locked vesting parameters
- * @returns The total token supply
- */
-export function getTotalTokenSupply(
-    swapBaseAmount: BN,
-    migrationBaseThreshold: BN,
-    lockedVestingParams: {
-        amountPerPeriod: BN
-        numberOfPeriod: BN
-        cliffUnlockAmount: BN
-    }
-): BN {
-    try {
-        // calculate total circulating amount
-        const totalCirculatingAmount = swapBaseAmount.add(
-            migrationBaseThreshold
-        )
-
-        // calculate total locked vesting amount
-        const totalLockedVestingAmount =
-            lockedVestingParams.cliffUnlockAmount.add(
-                lockedVestingParams.amountPerPeriod.mul(
-                    lockedVestingParams.numberOfPeriod
-                )
-            )
-
-        // calculate total amount
-        const totalAmount = totalCirculatingAmount.add(totalLockedVestingAmount)
-
-        // check for overflow
-        if (totalAmount.isNeg() || totalAmount.bitLength() > 64) {
-            throw new Error('Math overflow')
-        }
-
-        return totalAmount
-    } catch (error) {
-        throw new Error('Math overflow')
-    }
 }
 
 /**
