+++ packages/dynamic-bonding-curve/src/services/partner.ts
@@ -1,12 +1,12 @@
 import {
+    Commitment,
     SystemProgram,
     TransactionInstruction,
     type Connection,
     type Transaction,
 } from '@solana/web3.js'
-import type { DynamicBondingCurveProgramClient } from '../client'
+import { DynamicBondingCurveProgram } from './DbcProgram'
 import {
-    TokenType,
     type ClaimTradingFeeParam,
     type ConfigParameters,
     type CreateConfigParam,
@@ -18,25 +18,19 @@ import {
     BuildCurveAndCreateConfigParam,
 } from '../types'
 import {
-    deriveEventAuthority,
     derivePartnerMetadata,
-    derivePoolAuthority,
-} from '../derive'
-import {
-    createAssociatedTokenAccountIdempotentInstruction,
-    NATIVE_MINT,
-    TOKEN_2022_PROGRAM_ID,
-    TOKEN_PROGRAM_ID,
-} from '@solana/spl-token'
-import { findAssociatedTokenAddress, unwrapSOLInstruction } from '../utils'
-import { validateConfigParameters } from '../checks'
-import { buildCurve, buildCurveByMarketCap } from '../build'
+    unwrapSOLInstruction,
+    validateConfigParameters,
+    buildCurve,
+    buildCurveByMarketCap,
+    getTokenProgram,
+    getOrCreateATAInstruction,
+} from '../helpers'
+import { NATIVE_MINT } from '@solana/spl-token'
 
-export class PartnerService {
-    private connection: Connection
-
-    constructor(private programClient: DynamicBondingCurveProgramClient) {
-        this.connection = this.programClient.getProgram().provider.connection
+export class PartnerService extends DynamicBondingCurveProgram {
+    constructor(connection: Connection, commitment: Commitment) {
+        super(connection, commitment)
     }
 
     /**
@@ -47,8 +41,6 @@ export class PartnerService {
     async createConfig(
         createConfigParam: CreateConfigParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-
         const {
             config,
             feeClaimer,
@@ -58,24 +50,18 @@ export class PartnerService {
             ...configParam
         } = createConfigParam
 
-        const eventAuthority = deriveEventAuthority()
-
         // error checks
         validateConfigParameters({ ...configParam, leftoverReceiver })
 
-        const accounts = {
-            config,
-            feeClaimer,
-            leftoverReceiver,
-            quoteMint,
-            payer,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
+        return this.program.methods
             .createConfig(configParam)
-            .accounts(accounts)
+            .accountsPartial({
+                config,
+                feeClaimer,
+                leftoverReceiver,
+                quoteMint,
+                payer,
+            })
             .transaction()
     }
 
@@ -87,8 +73,6 @@ export class PartnerService {
     async buildCurveAndCreateConfig(
         buildCurveAndCreateConfigParam: BuildCurveAndCreateConfigParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-
         const {
             buildCurveParam,
             feeClaimer,
@@ -97,8 +81,6 @@ export class PartnerService {
             quoteMint,
             config,
         } = buildCurveAndCreateConfigParam
-
-        const eventAuthority = deriveEventAuthority()
 
         const curveConfig: ConfigParameters = buildCurve({
             ...buildCurveParam,
@@ -110,19 +92,15 @@ export class PartnerService {
             leftoverReceiver,
         })
 
-        const accounts = {
-            config,
-            feeClaimer,
-            leftoverReceiver,
-            quoteMint,
-            payer,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
+        return this.program.methods
             .createConfig(curveConfig)
-            .accounts(accounts)
+            .accounts({
+                config,
+                feeClaimer,
+                leftoverReceiver,
+                quoteMint,
+                payer,
+            })
             .transaction()
     }
 
@@ -134,8 +112,6 @@ export class PartnerService {
     async buildCurveAndCreateConfigByMarketCap(
         buildCurveAndCreateConfigByMarketCapParam: BuildCurveAndCreateConfigByMarketCapParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-
         const {
             buildCurveByMarketCapParam,
             feeClaimer,
@@ -144,8 +120,6 @@ export class PartnerService {
             quoteMint,
             config,
         } = buildCurveAndCreateConfigByMarketCapParam
-
-        const eventAuthority = deriveEventAuthority()
 
         const curveConfig: ConfigParameters = buildCurveByMarketCap({
             ...buildCurveByMarketCapParam,
@@ -157,19 +131,15 @@ export class PartnerService {
             leftoverReceiver,
         })
 
-        const accounts = {
-            config,
-            feeClaimer,
-            leftoverReceiver,
-            quoteMint,
-            payer,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
+        return this.program.methods
             .createConfig(curveConfig)
-            .accounts(accounts)
+            .accounts({
+                config,
+                feeClaimer,
+                leftoverReceiver,
+                quoteMint,
+                payer,
+            })
             .transaction()
     }
 
@@ -181,11 +151,9 @@ export class PartnerService {
     async createPartnerMetadata(
         createPartnerMetadataParam: CreatePartnerMetadataParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-        const eventAuthority = deriveEventAuthority()
         const partnerMetadata = derivePartnerMetadata(
             createPartnerMetadataParam.feeClaimer,
-            program.programId
+            this.program.programId
         )
 
         const partnerMetadataParam: CreatePartnerMetadataParameters = {
@@ -195,18 +163,14 @@ export class PartnerService {
             logo: createPartnerMetadataParam.logo,
         }
 
-        const accounts = {
-            partnerMetadata,
-            payer: createPartnerMetadataParam.payer,
-            feeClaimer: createPartnerMetadataParam.feeClaimer,
-            systemProgram: SystemProgram.programId,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
+        return this.program.methods
             .createPartnerMetadata(partnerMetadataParam)
-            .accounts(accounts)
+            .accountsPartial({
+                partnerMetadata,
+                payer: createPartnerMetadataParam.payer,
+                feeClaimer: createPartnerMetadataParam.feeClaimer,
+                systemProgram: SystemProgram.programId,
+            })
             .transaction()
     }
 
@@ -218,11 +182,7 @@ export class PartnerService {
     async claimTradingFee(
         claimTradingFeeParam: ClaimTradingFeeParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-        const poolAuthority = derivePoolAuthority(program.programId)
-        const eventAuthority = deriveEventAuthority()
-
-        const virtualPoolState = await this.programClient.getPool(
+        const virtualPoolState = await this.fetchVirtualPoolState(
             claimTradingFeeParam.pool
         )
 
@@ -232,94 +192,54 @@ export class PartnerService {
             )
         }
 
-        const poolConfigState = await this.programClient.getPoolConfig(
+        const poolConfigState = await this.fetchPoolConfigState(
             virtualPoolState.config
         )
 
-        const tokenBaseProgram =
-            poolConfigState.tokenType === TokenType.SPL
-                ? TOKEN_PROGRAM_ID
-                : TOKEN_2022_PROGRAM_ID
+        const tokenBaseProgram = getTokenProgram(poolConfigState.tokenType)
+        const tokenQuoteProgram = getTokenProgram(
+            poolConfigState.quoteTokenFlag
+        )
 
-        const tokenQuoteProgram =
-            poolConfigState.quoteTokenFlag === TokenType.SPL
-                ? TOKEN_PROGRAM_ID
-                : TOKEN_2022_PROGRAM_ID
-
-        const tokenBaseAccount = findAssociatedTokenAddress(
+        const postInstructions: TransactionInstruction[] = []
+        const {
+            ataTokenA: tokenBaseAccount,
+            ataTokenB: tokenQuoteAccount,
+            instructions: preInstructions,
+        } = await this.prepareTokenAccounts(
             claimTradingFeeParam.feeClaimer,
             virtualPoolState.baseMint,
-            tokenBaseProgram
-        )
-
-        const tokenQuoteAccount = findAssociatedTokenAddress(
-            claimTradingFeeParam.feeClaimer,
             poolConfigState.quoteMint,
+            tokenBaseProgram,
             tokenQuoteProgram
         )
-
-        const preInstructions: TransactionInstruction[] = []
-        const postInstructions: TransactionInstruction[] = []
-
-        const createBaseTokenAccountIx =
-            createAssociatedTokenAccountIdempotentInstruction(
-                claimTradingFeeParam.feeClaimer,
-                tokenBaseAccount,
-                claimTradingFeeParam.feeClaimer,
-                virtualPoolState.baseMint,
-                tokenBaseProgram
-            )
-
-        if (createBaseTokenAccountIx) {
-            preInstructions.push(createBaseTokenAccountIx)
-        }
-
-        const createQuoteTokenAccountIx =
-            createAssociatedTokenAccountIdempotentInstruction(
-                claimTradingFeeParam.feeClaimer,
-                tokenQuoteAccount,
-                claimTradingFeeParam.feeClaimer,
-                poolConfigState.quoteMint,
-                tokenQuoteProgram
-            )
-
-        if (createQuoteTokenAccountIx) {
-            preInstructions.push(createQuoteTokenAccountIx)
-        }
 
         if (poolConfigState.quoteMint.equals(NATIVE_MINT)) {
             const unwrapSolIx = unwrapSOLInstruction(
                 claimTradingFeeParam.feeClaimer
             )
-
-            if (unwrapSolIx) {
-                postInstructions.push(unwrapSolIx)
-            }
+            unwrapSolIx && postInstructions.push(unwrapSolIx)
         }
 
-        const accounts = {
-            poolAuthority,
-            config: virtualPoolState.config,
-            pool: claimTradingFeeParam.pool,
-            tokenAAccount: tokenBaseAccount,
-            tokenBAccount: tokenQuoteAccount,
-            baseVault: virtualPoolState.baseVault,
-            quoteVault: virtualPoolState.quoteVault,
-            baseMint: virtualPoolState.baseMint,
-            quoteMint: poolConfigState.quoteMint,
-            feeClaimer: claimTradingFeeParam.feeClaimer,
-            tokenBaseProgram,
-            tokenQuoteProgram,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
+        return this.program.methods
             .claimTradingFee(
                 claimTradingFeeParam.maxBaseAmount,
                 claimTradingFeeParam.maxQuoteAmount
             )
-            .accounts(accounts)
+            .accountsPartial({
+                poolAuthority: this.poolAuthority,
+                config: virtualPoolState.config,
+                pool: claimTradingFeeParam.pool,
+                tokenAAccount: tokenBaseAccount,
+                tokenBAccount: tokenQuoteAccount,
+                baseVault: virtualPoolState.baseVault,
+                quoteVault: virtualPoolState.quoteVault,
+                baseMint: virtualPoolState.baseMint,
+                quoteMint: poolConfigState.quoteMint,
+                feeClaimer: claimTradingFeeParam.feeClaimer,
+                tokenBaseProgram,
+                tokenQuoteProgram,
+            })
             .preInstructions(preInstructions)
             .postInstructions(postInstructions)
             .transaction()
@@ -333,11 +253,7 @@ export class PartnerService {
     async partnerWithdrawSurplus(
         partnerWithdrawSurplusParam: PartnerWithdrawSurplusParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-        const poolAuthority = derivePoolAuthority(program.programId)
-        const eventAuthority = deriveEventAuthority()
-
-        const virtualPoolState = await this.programClient.getPool(
+        const virtualPoolState = await this.fetchVirtualPoolState(
             partnerWithdrawSurplusParam.virtualPool
         )
         if (!virtualPoolState) {
@@ -346,63 +262,48 @@ export class PartnerService {
             )
         }
 
-        const poolConfigState = await this.programClient.getPoolConfig(
+        const poolConfigState = await this.fetchPoolConfigState(
             virtualPoolState.config
         )
 
-        const tokenQuoteProgram =
-            poolConfigState.quoteTokenFlag === TokenType.SPL
-                ? TOKEN_PROGRAM_ID
-                : TOKEN_2022_PROGRAM_ID
-
-        const tokenQuoteAccount = findAssociatedTokenAddress(
-            partnerWithdrawSurplusParam.feeClaimer,
-            poolConfigState.quoteMint,
-            tokenQuoteProgram
+        const tokenQuoteProgram = getTokenProgram(
+            poolConfigState.quoteTokenFlag
         )
 
         const preInstructions: TransactionInstruction[] = []
         const postInstructions: TransactionInstruction[] = []
 
-        const createQuoteTokenAccountIx =
-            createAssociatedTokenAccountIdempotentInstruction(
-                partnerWithdrawSurplusParam.feeClaimer,
-                tokenQuoteAccount,
-                partnerWithdrawSurplusParam.feeClaimer,
+        const { ataPubkey: tokenQuoteAccount, ix: createQuoteTokenAccountIx } =
+            await getOrCreateATAInstruction(
+                this.connection,
                 poolConfigState.quoteMint,
+                partnerWithdrawSurplusParam.feeClaimer,
+                partnerWithdrawSurplusParam.feeClaimer,
+                true,
                 tokenQuoteProgram
             )
 
-        if (createQuoteTokenAccountIx) {
+        createQuoteTokenAccountIx &&
             preInstructions.push(createQuoteTokenAccountIx)
-        }
 
         if (poolConfigState.quoteMint.equals(NATIVE_MINT)) {
             const unwrapSolIx = unwrapSOLInstruction(
                 partnerWithdrawSurplusParam.feeClaimer
             )
-
-            if (unwrapSolIx) {
-                postInstructions.push(unwrapSolIx)
-            }
+            unwrapSolIx && postInstructions.push(unwrapSolIx)
         }
-
-        const accounts = {
-            poolAuthority,
-            config: virtualPoolState.config,
-            virtualPool: partnerWithdrawSurplusParam.virtualPool,
-            tokenQuoteAccount,
-            quoteVault: virtualPoolState.quoteVault,
-            quoteMint: poolConfigState.quoteMint,
-            feeClaimer: partnerWithdrawSurplusParam.feeClaimer,
-            tokenQuoteProgram,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
+        return this.program.methods
             .partnerWithdrawSurplus()
-            .accounts(accounts)
+            .accountsPartial({
+                poolAuthority: this.poolAuthority,
+                config: virtualPoolState.config,
+                virtualPool: partnerWithdrawSurplusParam.virtualPool,
+                tokenQuoteAccount,
+                quoteVault: virtualPoolState.quoteVault,
+                quoteMint: poolConfigState.quoteMint,
+                feeClaimer: partnerWithdrawSurplusParam.feeClaimer,
+                tokenQuoteProgram,
+            })
             .preInstructions(preInstructions)
             .postInstructions(postInstructions)
             .transaction()
@@ -416,11 +317,7 @@ export class PartnerService {
     async withdrawLeftover(
         withdrawLeftoverParam: WithdrawLeftoverParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-        const poolAuthority = derivePoolAuthority(program.programId)
-        const eventAuthority = deriveEventAuthority()
-
-        const virtualPoolState = await this.programClient.getPool(
+        const virtualPoolState = await this.fetchVirtualPoolState(
             withdrawLeftoverParam.virtualPool
         )
 
@@ -430,52 +327,38 @@ export class PartnerService {
             )
         }
 
-        const poolConfigState = await this.programClient.getPoolConfig(
+        const poolConfigState = await this.fetchPoolConfigState(
             virtualPoolState.config
         )
 
-        const tokenBaseProgram =
-            poolConfigState.tokenType === TokenType.SPL
-                ? TOKEN_PROGRAM_ID
-                : TOKEN_2022_PROGRAM_ID
-
-        const tokenBaseAccount = findAssociatedTokenAddress(
-            poolConfigState.leftoverReceiver,
-            virtualPoolState.baseMint,
-            tokenBaseProgram
-        )
+        const tokenBaseProgram = getTokenProgram(poolConfigState.tokenType)
 
         const preInstructions: TransactionInstruction[] = []
-
-        const createBaseTokenAccountIx =
-            createAssociatedTokenAccountIdempotentInstruction(
-                poolConfigState.leftoverReceiver,
-                tokenBaseAccount,
-                poolConfigState.leftoverReceiver,
+        const { ataPubkey: tokenBaseAccount, ix: createBaseTokenAccountIx } =
+            await getOrCreateATAInstruction(
+                this.connection,
                 virtualPoolState.baseMint,
+                poolConfigState.leftoverReceiver,
+                poolConfigState.leftoverReceiver,
+                true,
                 tokenBaseProgram
             )
 
-        if (createBaseTokenAccountIx) {
+        createBaseTokenAccountIx &&
             preInstructions.push(createBaseTokenAccountIx)
-        }
 
-        const accounts = {
-            poolAuthority,
-            config: virtualPoolState.config,
-            virtualPool: withdrawLeftoverParam.virtualPool,
-            tokenBaseAccount,
-            baseVault: virtualPoolState.baseVault,
-            baseMint: virtualPoolState.baseMint,
-            leftoverReceiver: poolConfigState.leftoverReceiver,
-            tokenBaseProgram,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
+        return this.program.methods
             .withdrawLeftover()
-            .accounts(accounts)
+            .accountsPartial({
+                poolAuthority: this.poolAuthority,
+                config: virtualPoolState.config,
+                virtualPool: withdrawLeftoverParam.virtualPool,
+                tokenBaseAccount,
+                baseVault: virtualPoolState.baseVault,
+                baseMint: virtualPoolState.baseMint,
+                leftoverReceiver: poolConfigState.leftoverReceiver,
+                tokenBaseProgram,
+            })
             .preInstructions(preInstructions)
             .transaction()
     }
