+++ packages/dynamic-bonding-curve/src/services/pool.ts
@@ -1,51 +1,156 @@
 import {
+    Commitment,
     PublicKey,
     SystemProgram,
     TransactionInstruction,
     type Connection,
     type Transaction,
 } from '@solana/web3.js'
-import type { DynamicBondingCurveProgramClient } from '../client'
+import { DynamicBondingCurveProgram } from './DbcProgram'
 import {
+    InitializePoolBaseParam,
+    PrepareSwapParams,
     TokenType,
     type CreatePoolParam,
     type CreateVirtualPoolMetadataParam,
-    type CreateVirtualPoolMetadataParameters,
-    type InitializeVirtualPoolWithSplTokenAccounts,
-    type InitializeVirtualPoolWithToken2022Accounts,
-    type SwapAccounts,
     type SwapParam,
     type SwapQuoteParam,
 } from '../types'
 import {
-    deriveEventAuthority,
     deriveMetadata,
     derivePool,
-    derivePoolAuthority,
     deriveTokenVaultAddress,
     deriveVirtualPoolMetadata,
-} from '../derive'
-import {
-    createAssociatedTokenAccountIdempotentInstruction,
-    TOKEN_2022_PROGRAM_ID,
-} from '@solana/spl-token'
-import { TOKEN_PROGRAM_ID } from '@solana/spl-token'
-import { METAPLEX_PROGRAM_ID } from '../constants'
-import { prepareSwapParams } from '../common'
-import {
-    findAssociatedTokenAddress,
-    isNativeSol,
+    getTokenProgram,
     unwrapSOLInstruction,
     wrapSOLInstruction,
-} from '../utils'
+} from '../helpers'
+import { NATIVE_MINT, TOKEN_2022_PROGRAM_ID } from '@solana/spl-token'
+import { TOKEN_PROGRAM_ID } from '@solana/spl-token'
+import { METAPLEX_PROGRAM_ID } from '../constants'
 import { swapQuote } from '../math/swapQuote'
-import { validateBalance, validateBaseTokenType } from '../checks'
 
-export class PoolService {
-    private connection: Connection
+export class PoolService extends DynamicBondingCurveProgram {
+    constructor(connection: Connection, commitment: Commitment) {
+        super(connection, commitment)
+    }
 
-    constructor(private programClient: DynamicBondingCurveProgramClient) {
-        this.connection = this.programClient.getProgram().provider.connection
+    private async initializeSplPool(
+        initializeSplPoolParams: InitializePoolBaseParam
+    ): Promise<Transaction> {
+        const {
+            name,
+            symbol,
+            uri,
+            pool,
+            config,
+            payer,
+            poolCreator,
+            mintMetadata,
+            baseMint,
+            baseVault,
+            quoteVault,
+            quoteMint,
+        } = initializeSplPoolParams
+        return this.program.methods
+            .initializeVirtualPoolWithSplToken({
+                name,
+                symbol,
+                uri,
+            })
+            .accountsPartial({
+                pool,
+                config,
+                payer,
+                creator: poolCreator,
+                mintMetadata,
+                baseMint,
+                poolAuthority: this.poolAuthority,
+                baseVault,
+                quoteVault,
+                quoteMint,
+                tokenQuoteProgram: TOKEN_PROGRAM_ID,
+                metadataProgram: METAPLEX_PROGRAM_ID,
+                tokenProgram: TOKEN_PROGRAM_ID,
+            })
+            .transaction()
+    }
+
+    private async initializeToken2022Pool(
+        initializeToken2022PoolParams: InitializePoolBaseParam
+    ): Promise<Transaction> {
+        const {
+            name,
+            symbol,
+            uri,
+            pool,
+            config,
+            payer,
+            poolCreator,
+            baseMint,
+            baseVault,
+            quoteVault,
+            quoteMint,
+        } = initializeToken2022PoolParams
+        return this.program.methods
+            .initializeVirtualPoolWithToken2022({
+                name,
+                symbol,
+                uri,
+            })
+            .accountsPartial({
+                pool,
+                config,
+                payer,
+                creator: poolCreator,
+                baseMint,
+                poolAuthority: this.poolAuthority,
+                baseVault,
+                quoteVault,
+                quoteMint,
+                tokenQuoteProgram: TOKEN_PROGRAM_ID,
+                tokenProgram: TOKEN_2022_PROGRAM_ID,
+            })
+            .transaction()
+    }
+
+    /**
+     * Prepare swap parameters
+     * @param swapBaseForQuote - Whether to swap base for quote
+     * @param virtualPoolState - The virtual pool state
+     * @param poolConfigState - The pool config state
+     * @returns The prepare swap parameters
+     */
+    private prepareSwapParams(
+        swapBaseForQuote: boolean,
+        virtualPoolState: {
+            baseMint: PublicKey
+            poolType: TokenType
+        },
+        poolConfigState: {
+            quoteMint: PublicKey
+            quoteTokenFlag: TokenType
+        }
+    ): PrepareSwapParams {
+        if (swapBaseForQuote) {
+            return {
+                inputMint: new PublicKey(virtualPoolState.baseMint),
+                outputMint: new PublicKey(poolConfigState.quoteMint),
+                inputTokenProgram: getTokenProgram(virtualPoolState.poolType),
+                outputTokenProgram: getTokenProgram(
+                    poolConfigState.quoteTokenFlag
+                ),
+            }
+        } else {
+            return {
+                inputMint: new PublicKey(poolConfigState.quoteMint),
+                outputMint: new PublicKey(virtualPoolState.baseMint),
+                inputTokenProgram: getTokenProgram(
+                    poolConfigState.quoteTokenFlag
+                ),
+                outputTokenProgram: getTokenProgram(virtualPoolState.poolType),
+            }
+        }
     }
 
     /**
@@ -54,100 +159,50 @@ export class PoolService {
      * @returns A new pool
      */
     async createPool(createPoolParam: CreatePoolParam): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-        const {
+        const { baseMint, config, name, symbol, uri, payer, poolCreator } =
+            createPoolParam
+
+        const poolConfigState = await this.fetchPoolConfigState(config)
+
+        const { quoteMint, tokenType } = poolConfigState
+
+        const pool = derivePool(
             quoteMint,
             baseMint,
             config,
-            baseTokenType,
-            quoteTokenType,
-            name,
-            symbol,
-            uri,
-            payer,
-            poolCreator,
-        } = createPoolParam
-
-        const poolConfigState = await this.programClient.getPoolConfig(config)
-
-        // error checks
-        validateBaseTokenType(baseTokenType, poolConfigState)
-
-        const eventAuthority = deriveEventAuthority()
-        const poolAuthority = derivePoolAuthority(program.programId)
-        const pool = derivePool(quoteMint, baseMint, config, program.programId)
+            this.program.programId
+        )
         const baseVault = deriveTokenVaultAddress(
             pool,
             baseMint,
-            program.programId
+            this.program.programId
         )
         const quoteVault = deriveTokenVaultAddress(
             pool,
             quoteMint,
-            program.programId
+            this.program.programId
         )
-        const baseMetadata = deriveMetadata(baseMint)
 
-        if (baseTokenType === TokenType.SPL) {
-            const accounts: InitializeVirtualPoolWithSplTokenAccounts = {
-                pool,
-                config,
-                payer,
-                creator: poolCreator,
-                mintMetadata: baseMetadata,
-                program: program.programId,
-                tokenQuoteProgram:
-                    quoteTokenType === TokenType.SPL
-                        ? TOKEN_PROGRAM_ID
-                        : TOKEN_2022_PROGRAM_ID,
-                baseMint,
-                poolAuthority,
-                baseVault,
-                quoteVault,
-                quoteMint,
-                eventAuthority,
-                metadataProgram: METAPLEX_PROGRAM_ID,
-                systemProgram: SystemProgram.programId,
-                tokenProgram: TOKEN_PROGRAM_ID,
-            }
-            return program.methods
-                .initializeVirtualPoolWithSplToken({
-                    name,
-                    symbol,
-                    uri,
-                })
-                .accounts(accounts)
-                .transaction()
+        const baseParams: InitializePoolBaseParam = {
+            name,
+            symbol,
+            uri,
+            pool,
+            config,
+            payer,
+            poolCreator,
+            baseMint,
+            baseVault,
+            quoteVault,
+            quoteMint,
         }
 
-        if (baseTokenType === TokenType.Token2022) {
-            const accounts: InitializeVirtualPoolWithToken2022Accounts = {
-                pool,
-                config,
-                payer,
-                creator: poolCreator,
-                program: program.programId,
-                baseMint,
-                poolAuthority,
-                baseVault,
-                quoteVault,
-                quoteMint,
-                eventAuthority,
-                tokenQuoteProgram: TOKEN_PROGRAM_ID,
-                tokenProgram: TOKEN_2022_PROGRAM_ID,
-                systemProgram: SystemProgram.programId,
-            }
-            return program.methods
-                .initializeVirtualPoolWithToken2022({
-                    name,
-                    symbol,
-                    uri,
-                })
-                .accounts(accounts)
-                .transaction()
+        if (tokenType === TokenType.SPL) {
+            const mintMetadata = deriveMetadata(baseMint)
+            return this.initializeSplPool({ ...baseParams, mintMetadata })
+        } else {
+            return this.initializeToken2022Pool(baseParams)
         }
-
-        throw new Error('Invalid base token type')
     }
 
     /**
@@ -158,32 +213,23 @@ export class PoolService {
     async createPoolMetadata(
         createVirtualPoolMetadataParam: CreateVirtualPoolMetadataParam
     ): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-        const eventAuthority = deriveEventAuthority()
         const virtualPoolMetadata = deriveVirtualPoolMetadata(
             createVirtualPoolMetadataParam.virtualPool
         )
-
-        const virtualPoolMetadataParam: CreateVirtualPoolMetadataParameters = {
-            padding: new Array(96).fill(0),
-            name: createVirtualPoolMetadataParam.name,
-            website: createVirtualPoolMetadataParam.website,
-            logo: createVirtualPoolMetadataParam.logo,
-        }
-
-        const accounts = {
-            virtualPool: createVirtualPoolMetadataParam.virtualPool,
-            virtualPoolMetadata,
-            creator: createVirtualPoolMetadataParam.creator,
-            payer: createVirtualPoolMetadataParam.payer,
-            systemProgram: SystemProgram.programId,
-            eventAuthority,
-            program: program.programId,
-        }
-
-        return program.methods
-            .createVirtualPoolMetadata(virtualPoolMetadataParam)
-            .accounts(accounts)
+        return this.program.methods
+            .createVirtualPoolMetadata({
+                padding: new Array(96).fill(0),
+                name: createVirtualPoolMetadataParam.name,
+                website: createVirtualPoolMetadataParam.website,
+                logo: createVirtualPoolMetadataParam.logo,
+            })
+            .accountsPartial({
+                virtualPool: createVirtualPoolMetadataParam.virtualPool,
+                virtualPoolMetadata,
+                creator: createVirtualPoolMetadataParam.creator,
+                payer: createVirtualPoolMetadataParam.payer,
+                systemProgram: SystemProgram.programId,
+            })
             .transaction()
     }
 
@@ -194,17 +240,13 @@ export class PoolService {
      * @returns A swap transaction
      */
     async swap(pool: PublicKey, swapParam: SwapParam): Promise<Transaction> {
-        const program = this.programClient.getProgram()
-        const eventAuthority = deriveEventAuthority()
-        const poolAuthority = derivePoolAuthority(program.programId)
-
-        const virtualPoolState = await this.programClient.getPool(pool)
+        const virtualPoolState = await this.fetchVirtualPoolState(pool)
 
         if (!virtualPoolState) {
             throw new Error(`Pool not found: ${pool.toString()}`)
         }
 
-        const poolConfigState = await this.programClient.getPoolConfig(
+        const poolConfigState = await this.fetchPoolConfigState(
             virtualPoolState.config
         )
 
@@ -212,91 +254,27 @@ export class PoolService {
             swapParam
 
         const { inputMint, outputMint, inputTokenProgram, outputTokenProgram } =
-            prepareSwapParams(
+            this.prepareSwapParams(
                 swapBaseForQuote,
                 virtualPoolState,
                 poolConfigState
             )
 
-        const isSOLInput = isNativeSol(inputMint)
-        const isSOLOutput = isNativeSol(outputMint)
-
-        const inputTokenAccount = findAssociatedTokenAddress(
+        // Add preInstructions for ATA creation and SOL wrapping
+        const {
+            ataTokenA: inputTokenAccount,
+            ataTokenB: outputTokenAccount,
+            instructions: preInstructions,
+        } = await this.prepareTokenAccounts(
             owner,
             inputMint,
-            inputTokenProgram
-        )
-
-        const outputTokenAccount = findAssociatedTokenAddress(
-            owner,
             outputMint,
+            inputTokenProgram,
             outputTokenProgram
         )
 
-        await validateBalance(
-            this.connection,
-            owner,
-            inputMint,
-            amountIn,
-            inputTokenAccount
-        )
-
-        const accounts: SwapAccounts = {
-            baseMint: virtualPoolState.baseMint,
-            quoteMint: poolConfigState.quoteMint,
-            pool: pool,
-            baseVault: virtualPoolState.baseVault,
-            quoteVault: virtualPoolState.quoteVault,
-            config: virtualPoolState.config,
-            eventAuthority,
-            poolAuthority,
-            referralTokenAccount: null,
-            inputTokenAccount,
-            outputTokenAccount,
-            payer: owner,
-            tokenBaseProgram: swapBaseForQuote
-                ? inputTokenProgram
-                : outputTokenProgram,
-            tokenQuoteProgram: swapBaseForQuote
-                ? outputTokenProgram
-                : inputTokenProgram,
-            program: program.programId,
-        }
-
-        // Add preInstructions for ATA creation and SOL wrapping
-        const preInstructions: TransactionInstruction[] = []
-
-        // Check and create ATAs if needed
-        const inputTokenAccountInfo =
-            await this.connection.getAccountInfo(inputTokenAccount)
-        if (!inputTokenAccountInfo) {
-            preInstructions.push(
-                createAssociatedTokenAccountIdempotentInstruction(
-                    owner,
-                    inputTokenAccount,
-                    owner,
-                    inputMint,
-                    inputTokenProgram
-                )
-            )
-        }
-
-        const outputTokenAccountInfo =
-            await this.connection.getAccountInfo(outputTokenAccount)
-        if (!outputTokenAccountInfo) {
-            preInstructions.push(
-                createAssociatedTokenAccountIdempotentInstruction(
-                    owner,
-                    outputTokenAccount,
-                    owner,
-                    outputMint,
-                    outputTokenProgram
-                )
-            )
-        }
-
         // Add SOL wrapping instructions if needed
-        if (isSOLInput) {
+        if (inputMint.equals(NATIVE_MINT)) {
             preInstructions.push(
                 ...wrapSOLInstruction(
                     owner,
@@ -308,19 +286,40 @@ export class PoolService {
 
         // Add postInstructions for SOL unwrapping
         const postInstructions: TransactionInstruction[] = []
-        if (isSOLInput || isSOLOutput) {
+        if (
+            [inputMint.toBase58(), outputMint.toBase58()].includes(
+                NATIVE_MINT.toBase58()
+            )
+        ) {
             const unwrapIx = unwrapSOLInstruction(owner)
-            if (unwrapIx) {
-                postInstructions.push(unwrapIx)
-            }
+
+            unwrapIx && postInstructions.push(unwrapIx)
         }
 
-        return program.methods
+        return this.program.methods
             .swap({
                 amountIn,
                 minimumAmountOut,
             })
-            .accounts(accounts)
+            .accountsPartial({
+                baseMint: virtualPoolState.baseMint,
+                quoteMint: poolConfigState.quoteMint,
+                pool: pool,
+                baseVault: virtualPoolState.baseVault,
+                quoteVault: virtualPoolState.quoteVault,
+                config: virtualPoolState.config,
+                poolAuthority: this.poolAuthority,
+                referralTokenAccount: null,
+                inputTokenAccount,
+                outputTokenAccount,
+                payer: owner,
+                tokenBaseProgram: swapBaseForQuote
+                    ? inputTokenProgram
+                    : outputTokenProgram,
+                tokenQuoteProgram: swapBaseForQuote
+                    ? outputTokenProgram
+                    : inputTokenProgram,
+            })
             .preInstructions(preInstructions)
             .postInstructions(postInstructions)
             .transaction()
