+++ packages/dynamic-bonding-curve/src/types.ts
@@ -304,6 +304,21 @@ export type CreatePartnerMetadataParam = {
     payer: PublicKey
 }
 
+export type InitializePoolBaseParam = {
+    name: string
+    symbol: string
+    uri: string
+    pool: PublicKey
+    config: PublicKey
+    payer: PublicKey
+    poolCreator: PublicKey
+    baseMint: PublicKey
+    baseVault: PublicKey
+    quoteVault: PublicKey
+    quoteMint: PublicKey
+    mintMetadata?: PublicKey
+}
+
 export type CreatePoolParam = {
     name: string
     symbol: string
@@ -312,9 +327,6 @@ export type CreatePoolParam = {
     poolCreator: PublicKey
     config: PublicKey
     baseMint: PublicKey
-    baseTokenType: TokenType
-    quoteMint: PublicKey
-    quoteTokenType: TokenType
 }
 
 export type SwapParam = {
